implement the new Router
this router integrates the rewritten trie implementation (which is now
mainly a tree of hashmaps, but keeps the same API).
It now supports matching domains and paths with regular expressions.
There are also pre and post matching rules than can be tested before and
after the tree.
The idea is that if you do not have many rules, you can put them in pre
or post and quickly try them one by one. But if you handle a lot of domains,
the tree implementation will be lighter, and allow testing less rules to
find the matching application